# Conflicts:
#	Developpement/Web/web_project/src/components/RichTextEditor.js
#	Developpement/Web/web_project/src/utils.js

+

URL UNIQUE ALBUM

+++ Developpement/Api/api_project/models/expense.js
@@ -10,7 +10,7 @@ module.exports = sequelize => {
     }
 
     Expense.init({
-        cost: DataTypes.INTEGER,
+        cost: DataTypes.DOUBLE,
         to: DataTypes.STRING,
         category: DataTypes.STRING,
         date: DataTypes.DATE
